SMV source: make changes to remove INTEL compiler warnings
Former-commit-id: 7857dbffe30bc9c5bb70c1d5bebab9a301b641e4 [formerly eb3122f997db11b2ec2e8d578de176b71398e871]
Former-commit-id: 584591bdedc883bac16c5829eac7ece044b416af
Former-commit-id: 58d93cd7ebfc3507abe2cf196fdd9184cf6f1503

+++ sv_5/source/smokeview/output.c
@@ -159,8 +159,8 @@ void bench_out(float localframerate){
   fprintf(fileout,"     Window height:%i\n",screenHeight);
 
 #ifdef pp_memstatus
-  fprintf(fileout,"       Memory Used:%iMB\n",memused);
-  fprintf(fileout,"      Total Memory:%iMB\n",tmem);
+  fprintf(fileout,"       Memory Used:%iMB\n",(int)memused);
+  fprintf(fileout,"      Total Memory:%iMB\n",(int)tmem);
 #else
   fprintf(fileout,"       Memory Used:\n");
   fprintf(fileout,"      Total Memory:\n");
